src/VumbaSoft.AdventureWorks.Controllers/HumanResources/Departments/DepartmentsController.cs
using Microsoft.AspNetCore.Mvc;
using VumbaSoft.AdventureWorks.Objects;
using VumbaSoft.AdventureWorks.Services;
using VumbaSoft.AdventureWorks.Validators;
using System;

namespace VumbaSoft.AdventureWorks.Controllers.HumanResources
{
    [Area("HumanResources")]
    public class DepartmentsController : ValidatedController<IDepartmentValidator, IDepartmentService>
    {
        public DepartmentsController(IDepartmentValidator validator, IDepartmentService service)
            : base(validator, service)
        {
        }

        [HttpGet]
        public ViewResult Index()
        {
            return View(Service.GetViews());
        }

        [HttpGet]
        public ViewResult Create()
        {
            return View();
        }

        [HttpPost]
        public ActionResult Create(DepartmentView department)
        {
            if (!Validator.CanCreate(department))
                return View(department);

            Service.Create(department);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Details(Int32 id)
        {
            return NotEmptyView(Service.Get<DepartmentView>(id));
        }

        [HttpGet]
        public ActionResult Edit(Int32 id)
        {
            return NotEmptyView(Service.Get<DepartmentView>(id));
        }

        [HttpPost]
        public ActionResult Edit(DepartmentView department)
        {
            if (!Validator.CanEdit(department))
                return View(department);

            Service.Edit(department);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Delete(Int32 id)
        {
            return NotEmptyView(Service.Get<DepartmentView>(id));
        }

        [HttpPost]
        [ActionName("Delete")]
        public RedirectToActionResult DeleteConfirmed(Int32 id)
        {
            Service.Delete(id);

            return RedirectToAction("Index");
        }
    }
}
